pyproject.toml
[project]
name = "football-match-api"
version = "0.2.0"
description = "Football match API"
authors = [
    {name = "Kleveta", email = "kleveta.lukas99@gmail.com"},
]
dependencies = [
    "uvicorn>=0.24.0.post1",
    "fastapi>=0.104.1",
    "pymongo>=4.6.1",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.1.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload"
